@@ -2,19 +2,22 @@
   "name": "docsify",
   "version": "4.13.0",
   "description": "A magical documentation generator.",
-  "author": {
-    "name": "qingwei-li",
-    "email": "cinwell.li@gmail.com",
-    "url": "https://github.com/QingWei-Li"
-  },
   "homepage": "https://docsify.js.org",
+  "repository": "github:docsifyjs/docsify",
+  "authors": "https://github.com/docsifyjs/docsify/graphs/contributors",
   "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/docsifyjs/docsify.git"
+  "collective": {
+    "url": "https://opencollective.com/docsify"
   },
+  "type": "module",
+  "// The 'main' and 'unpkg' fields will remain as legacy for backwards compatbility for now. We will add deprectaion warnings to these outputs to give people time to see the warnings in their apps in a non-breaking way, and eventually we can remove the legacy stuff.": "",
   "main": "lib/docsify.js",
   "unpkg": "lib/docsify.min.js",
+  "// We're using the 'exports' field as an override of the 'main' field to provide the new ESM setup. Once we remove legacy 'main', we will remove the 'exports' field and have a simple ESM setup with only a 'main' field.": "",
+  "exports": {
+    ".": "./src/Docsify.js",
+    "./*": "./*"
+  },
   "files": [
     "lib",
     "themes"
@@ -46,9 +49,10 @@
     "pub": "sh build/release.sh",
     "serve": "node server",
     "test:e2e": "playwright test",
-    "test:integration": "jest --selectProjects integration",
-    "test:unit": "jest --selectProjects unit",
-    "test": "jest && run-s test:e2e",
+    "test:integration": "npm run jest -- --selectProjects integration",
+    "test:unit": "npm run jest -- --selectProjects unit",
+    "test": "npm run jest && run-s test:e2e",
+    "jest": "cross-env NODE_OPTIONS=--experimental-vm-modules jest",
     "watch:css": "npm run css -- -o themes -w",
     "watch:js": "node build/build.js"
   },
@@ -73,7 +77,9 @@
     "@babel/core": "^7.11.6",
     "@babel/eslint-parser": "^7.16.5",
     "@babel/preset-env": "^7.11.5",
+    "@eslint/js": "^8.43.0",
     "@playwright/test": "^1.18.1",
+    "@types/eslint": "^8.40.2",
     "autoprefixer-stylus": "^1.0.0",
     "axios": "^0.21.1",
     "babel-jest": "^27.4.6",
@@ -84,19 +90,20 @@
     "copy-dir": "^1.2.0",
     "cross-env": "^7.0.3",
     "cssnano": "^4.1.10",
-    "eslint": "^8.7.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-import": "^2.20.1",
-    "eslint-plugin-jest": "^27.2.1",
-    "eslint-plugin-playwright": "^0.12.0",
-    "eslint-plugin-prettier": "^4.0.0",
+    "eslint": "^8.43.0",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-jest": "^27.2.2",
+    "eslint-plugin-playwright": "^0.15.1",
+    "eslint-plugin-prettier": "^4.2.1",
+    "globals": "^13.20.0",
     "husky": "^8.0.3",
     "jest": "^27.4.7",
     "lint-staged": "^13.2.2",
     "live-server": "^1.2.1",
     "mkdirp": "^3.0.0",
     "npm-run-all": "^4.1.5",
-    "prettier": "^2.5.1",
+    "prettier": "^2.8.8",
     "rimraf": "^3.0.0",
     "rollup": "^1.23.1",
     "rollup-plugin-async": "^1.2.0",
@@ -117,8 +124,5 @@
     "documentation",
     "creator",
     "generator"
-  ],
-  "collective": {
-    "url": "https://opencollective.com/docsify"
-  }
+  ]
 }
